Finaliza cambio de estructura y conexion a base de datos
