--- BCIDesafio-main/src/main/java/com/example/DesafioBCI/exception/UserAlreadyExistsException.java
@@ -1,10 +1,0 @@
-package com.example.DesafioBCI.exception;
-
-public class UserAlreadyExistsException  extends RuntimeException {
-
-	private static final long serialVersionUID = 1L;
-
-	public UserAlreadyExistsException(String message) {
-        super(message);
-    }
-}
